+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for OCRRefLangCompiler
 
+## 0.1.1.1 -- 2025-01-18
+* Added dependency on `System.Console.ANSI` to control colour of error messages
+* Added `Logging`module for futureproofing
+
 ## 0.1.1.0 -- 2025-01-18
 
 * Added preprocessing